Worked more with trying to record head bobs

+++ app/src/main/java/com/example/ben/headbobhero/HeadBob.java
@@ -10,9 +10,9 @@ import android.graphics.Point;
 class HeadBob {
     public HeadBobDirection direction;
 
-    public int offset;
+    public long offset;
 
-    public HeadBob(int offset, HeadBobDirection direction) {
+    public HeadBob(long offset, HeadBobDirection direction) {
         this.direction = direction;
         this.offset = offset;
     }
